monolithic_agent/requirements.txt
# AlgoAgent - Main Requirements File
# Install all dependencies: pip install -r requirements.txt

# Django Framework
Django>=4.2.0,<5.0
djangorestframework>=3.14.0

# Django Extensions
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
channels>=4.0.0  # WebSocket support
channels-redis>=4.1.0  # Redis channel layer (optional but recommended)
daphne>=4.0.0  # ASGI server for Channels

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter (optional, SQLite is default)

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# API & Web
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Validation & Schema
jsonschema>=4.20.0

# Environment Variables
python-dotenv>=1.0.0

# AI/ML Integration
google-generativeai>=0.3.0

# LangChain for conversation memory and context management
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-mock>=3.12.0

# Development Tools (optional)
ipython>=8.12.0
black>=23.0.0  # Code formatter
flake8>=6.0.0  # Linter

# Production Server (optional)
gunicorn>=21.0.0
whitenoise>=6.5.0  # Static file serving
